fix trackers; integrate empty layout ,loader and receiver; fix some logics

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/draft/presentation/fragment/AddEditProductDraftFragment.kt
@@ -1,24 +1,38 @@
 package com.tokopedia.product.addedit.draft.presentation.fragment
 
+import android.app.Activity
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
 import android.os.Bundle
 import android.view.*
-import android.widget.Toast
+import android.widget.Button
+import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
+import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.tokopedia.abstraction.base.view.fragment.BaseDaggerFragment
+import com.tokopedia.abstraction.common.utils.view.MethodChecker
+import com.tokopedia.globalerror.GlobalError
 import com.tokopedia.header.HeaderUnify
+import com.tokopedia.kotlin.extensions.view.hide
 import com.tokopedia.kotlin.extensions.view.observe
+import com.tokopedia.kotlin.extensions.view.show
 import com.tokopedia.product.addedit.R
+import com.tokopedia.product.addedit.common.constant.AddEditProductConstants.BROADCAST_ADD_PRODUCT
+import com.tokopedia.product.addedit.common.util.AddEditProductErrorHandler
 import com.tokopedia.product.addedit.draft.di.AddEditProductDraftComponent
-import com.tokopedia.product.addedit.draft.mapper.AddEditProductMapper
 import com.tokopedia.product.addedit.draft.presentation.adapter.ProductDraftListAdapter
 import com.tokopedia.product.addedit.draft.presentation.listener.ProductDraftListListener
-import com.tokopedia.product.addedit.draft.presentation.model.ProductDraftUiModel
 import com.tokopedia.product.addedit.draft.presentation.viewmodel.AddEditProductDraftViewModel
 import com.tokopedia.product.addedit.preview.presentation.activity.AddEditProductPreviewActivity
-import com.tokopedia.product.manage.common.draft.data.model.ProductDraft
+import com.tokopedia.product.addedit.tracking.ProductDraftTracking
 import com.tokopedia.unifycomponents.Toaster
 import com.tokopedia.usecase.coroutines.Fail
 import com.tokopedia.usecase.coroutines.Success
+import com.tokopedia.user.session.UserSession
+import com.tokopedia.user.session.UserSessionInterface
 import kotlinx.android.synthetic.main.fragment_add_edit_product_draft.*
 import javax.inject.Inject
 
@@ -26,6 +40,8 @@ class AddEditProductDraftFragment : BaseDaggerFragment(), ProductDraftListListen
 
     companion object {
         const val SCREEN_NAME = "/draft product page"
+        const val SCREEN_NAME_ADD_PRODUCT = "/add-product"
+        const val REQUEST_CODE_ADD_PRODUCT = 9003
 
         @JvmStatic
         fun newInstance() = AddEditProductDraftFragment()
@@ -34,14 +50,31 @@ class AddEditProductDraftFragment : BaseDaggerFragment(), ProductDraftListListen
     @Inject
     lateinit var viewModel: AddEditProductDraftViewModel
 
+    private var broadcastReceiver: BroadcastReceiver? = null
+    private var tvEmptyTitle: TextView? = null
+    private var tvEmptyContent: TextView? = null
+    private var btnAddProduct: Button? = null
+    private var userSession: UserSessionInterface? = null
     private var draftListAdapter: ProductDraftListAdapter? = null
-    private var drafts = mutableListOf<ProductDraftUiModel>()
+    private var mMenu: Menu? = null
     private var deletePosition = -1
+    private var draftListChanged = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setupToolbarActions()
         setHasOptionsMenu(true)
+        userSession = UserSession(context)
+    }
+
+    override fun onResume() {
+        super.onResume()
+        registerDraftReceiver()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        unregisterDraftReceiver()
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -55,18 +88,41 @@ class AddEditProductDraftFragment : BaseDaggerFragment(), ProductDraftListListen
     }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        mMenu = menu
         inflater.inflate(R.menu.menu_product_draft, menu)
         super.onCreateOptionsMenu(menu, inflater)
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        when(requestCode) {
+            REQUEST_CODE_ADD_PRODUCT -> {
+                if (resultCode == Activity.RESULT_OK) {
+                    ProductDraftTracking.sendScreenProductDraft(screenName, SCREEN_NAME_ADD_PRODUCT)
+                }
+            }
+        }
+        super.onActivityResult(requestCode, resultCode, data)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when(item.itemId) {
             R.id.item_label_add_product -> {
                 val intent = AddEditProductPreviewActivity.createInstance(context)
-                startActivity(intent)
+                startActivityForResult(intent, REQUEST_CODE_ADD_PRODUCT)
+                userSession?.shopId?.run {
+                    ProductDraftTracking.sendAddProductClick(this, ProductDraftTracking.CLICK_ADD_PRODUCT_WITHOUT_DRAFT)
+                }
             }
             R.id.item_delete_all_draft -> {
-                viewModel.deleteAllProductDraft()
+                val alertDialogBuilder = AlertDialog.Builder(requireContext())
+                        .setMessage(getString(R.string.label_draft_menu_delete_all_draft_dialog_message))
+                        .setNegativeButton(getString(R.string.label_cancel_button_on_dialog)) { _, _ -> }
+                        .setPositiveButton(getString(R.string.label_delete_button_on_dialog)) { _, _ ->
+                            viewModel.deleteAllProductDraft()
+                            displayLoader()
+                        }
+                val dialog = alertDialogBuilder.create()
+                dialog.show()
             }
         }
         return super.onOptionsItemSelected(item)
@@ -80,12 +136,27 @@ class AddEditProductDraftFragment : BaseDaggerFragment(), ProductDraftListListen
 
     override fun onDraftClickListener(draftId: Long) {
         val intent = AddEditProductPreviewActivity.createInstance(context, draftId.toString())
-        startActivity(intent)
+        startActivityForResult(intent, REQUEST_CODE_ADD_PRODUCT)
+        ProductDraftTracking.sendProductDraftClick(ProductDraftTracking.EDIT_DRAFT)
     }
 
-    override fun onDraftDeleteListener(draftId: Long, position: Int) {
-        viewModel.deleteProductDraft(draftId)
-        deletePosition = position
+    override fun onDraftDeleteListener(draftId: Long, position: Int, productName: String) {
+        val message = if (productName.isBlank()) {
+            getString(R.string.label_draft_delete_draft_dialog_message)
+        } else {
+            getString(R.string.label_draft_delete_draft_dialog_has_product_name_message, productName)
+        }
+        val alertDialogBuilder = AlertDialog.Builder(requireContext())
+                .setMessage(MethodChecker.fromHtml(message))
+                .setNegativeButton(getString(R.string.label_cancel)) { _, _ -> }
+                .setPositiveButton(getString(R.string.label_delete_button_on_dialog)) { _, _ ->
+                    viewModel.deleteProductDraft(draftId)
+                    deletePosition = position
+                    ProductDraftTracking.sendProductDraftClick(ProductDraftTracking.DELETE_DRAFT)
+                    displayLoader()
+                }
+        val dialog = alertDialogBuilder.create()
+        dialog.show()
     }
 
     private fun setupToolbarActions() {
@@ -98,33 +169,146 @@ class AddEditProductDraftFragment : BaseDaggerFragment(), ProductDraftListListen
     }
 
     private fun setup() {
-        draftListAdapter = ProductDraftListAdapter(this, drafts)
+        tvEmptyTitle = activity?.findViewById(com.tokopedia.baselist.R.id.text_view_empty_title_text)
+        tvEmptyContent = activity?.findViewById(com.tokopedia.baselist.R.id.text_view_empty_content_text)
+        btnAddProduct = activity?.findViewById(com.tokopedia.baselist.R.id.button_add_promo)
+
+        tvEmptyTitle?.text = getString(R.string.label_draft_product_empty)
+        tvEmptyContent?.hide()
+        btnAddProduct?.apply {
+            text = getString(R.string.label_draft_product_add_product)
+            setOnClickListener {
+                val intent = AddEditProductPreviewActivity.createInstance(context)
+                startActivityForResult(intent, REQUEST_CODE_ADD_PRODUCT)
+                ProductDraftTracking.sendProductDraftClick(ProductDraftTracking.ADD_PRODUCT)
+                userSession?.shopId?.run {
+                    ProductDraftTracking.sendAddProductClick(this, ProductDraftTracking.CLICK_ADD_PRODUCT)
+                }
+            }
+        }
+
+        draftListAdapter = ProductDraftListAdapter(this)
         rvDraft.adapter = draftListAdapter
         rvDraft.layoutManager = LinearLayoutManager(context)
 
         viewModel.getAllProductDraft()
+        displayLoader()
     }
 
     private fun observer() {
         observe(viewModel.drafts) {
+            dismissLoader()
             when(it) {
-                is Success -> draftListAdapter?.setDrafts(it.data)
-                is Fail -> {}
+                is Success -> {
+                    geDraft.hide()
+                    draftListAdapter?.setDrafts(it.data)
+                    displayEmptyListLayout()
+                }
+                is Fail -> {
+                    rvDraft.hide()
+                    geDraft.apply {
+                        setType(GlobalError.SERVER_ERROR)
+                        setActionClickListener {
+                            viewModel.getAllProductDraft()
+                        }
+                    }.show()
+                    AddEditProductErrorHandler.logExceptionToCrashlytics(it.throwable)
+                }
             }
         }
 
         observe(viewModel.deleteDraft) {
+            dismissLoader()
             when(it) {
-                is Success -> draftListAdapter?.deleteDraft(deletePosition)
-                is Fail -> {}
+                is Success -> {
+                    draftListAdapter?.deleteDraft(deletePosition)
+                    displayEmptyListLayout()
+                }
+                is Fail -> {
+                    AddEditProductErrorHandler.logExceptionToCrashlytics(it.throwable)
+                }
             }
         }
 
         observe(viewModel.deleteAllDraft) {
+            dismissLoader()
             when(it) {
-                is Success -> draftListAdapter?.deleteAllDrafts()
-                is Fail -> {}
+                is Success -> {
+                    Toaster.build(
+                            frameLayout,
+                            getString(R.string.label_draft_success_delete_draft_message),
+                            Toaster.LENGTH_SHORT,
+                            Toaster.TYPE_NORMAL
+                    ).show()
+                    draftListAdapter?.deleteAllDrafts()
+                    displayEmptyListLayout()
+                }
+                is Fail -> {
+                    Toaster.build(
+                            frameLayout,
+                            getString(R.string.label_draft_error_delete_draft_message),
+                            Toaster.LENGTH_SHORT,
+                            Toaster.TYPE_ERROR
+                    ).show()
+                    AddEditProductErrorHandler.logExceptionToCrashlytics(it.throwable)
+                }
             }
         }
+    }
+
+    private fun displayLoader() {
+        loaderUnify.show()
+        rvDraft.hide()
+    }
+
+    private fun dismissLoader() {
+        loaderUnify.hide()
+        rvDraft.show()
+    }
+
+    private fun displayEmptyListLayout() {
+        draftListAdapter?.isDraftEmpty()?.let { isEmpty ->
+            if (isEmpty) {
+                emptyLayout.show()
+                rvDraft.hide()
+                mMenu?.findItem(R.id.item_delete_all_draft)?.isVisible = false
+            } else {
+                rvDraft.show()
+                emptyLayout.hide()
+                mMenu?.findItem(R.id.item_delete_all_draft)?.isVisible = true
+            }
+        }
+    }
+
+    private fun registerDraftReceiver() {
+        broadcastReceiver = object : BroadcastReceiver() {
+            override fun onReceive(context: Context, intent: Intent) {
+                if (intent.action == BROADCAST_ADD_PRODUCT) {
+                    viewModel.getAllProductDraft()
+                    draftListChanged = true;
+                }
+            }
+        }
+
+        activity?.let {
+            val intentFilters = IntentFilter().apply {
+                addAction(BROADCAST_ADD_PRODUCT)
+            }
+            broadcastReceiver?.let { receiver ->
+                LocalBroadcastManager.getInstance(it).registerReceiver(receiver, intentFilters)
+            }
+        }
+    }
+
+    private fun unregisterDraftReceiver() {
+        activity?.let {
+            broadcastReceiver?.let { receiver ->
+                LocalBroadcastManager.getInstance(it).unregisterReceiver(receiver)
+            }
+        }
+    }
+
+    fun getDraftListChanged(): Boolean {
+        return draftListChanged
     }
 }
